removes context dependency from view builder

+++ app/src/main/java/com/atherton/darren/widget/presenter/AppRatingPresenter.java
@@ -17,5 +17,5 @@ public interface AppRatingPresenter {
 
     void onBackdropTouched();
 
-    void onRatingChanged(int rating);
+    void onRatingChanged(float rating);
 }
